requirements.txt
anyio==3.7.0
asgiref==3.7.2
backoff==2.2.1
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.4
coloredlogs==15.0.1
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.3.0
duckdb==0.8.1
exceptiongroup==1.1.2
fastapi==0.99.1
flatbuffers==23.5.26
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
lz4==4.3.2
mongoengine==0.27.0
monotonic==1.6
mpmath==1.3.0
numpy==1.25.0
onnxruntime==1.15.1
overrides==7.3.1
packaging==23.1
pandas==2.0.3
posthog==3.0.1
protobuf==4.23.3
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.10
PyJWT==2.7.0
pymongo==3.12.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tokenizers==0.13.3
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
zstandard==0.21.0
